init store quota and transaction

+++ src/main/java/com/xpay/pay/rest/contract/CreateStoreRequest.java
@@ -2,7 +2,7 @@ package com.xpay.pay.rest.contract;
 
 public class CreateStoreRequest {
 	private String name;
-	private int bailPercentage;
+	private Float bailPercentage;
 	private long appId;
 	private String csrTel;
 	private String proxyUrl;
@@ -12,10 +12,10 @@ public class CreateStoreRequest {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getBailPercentage() {
+	public Float getBailPercentage() {
 		return bailPercentage;
 	}
-	public void setBailPercentage(int bailPercentage) {
+	public void setBailPercentage(Float bailPercentage) {
 		this.bailPercentage = bailPercentage;
 	}
 	public long getAppId() {
