+++ src/main/java/com/xpay/pay/rest/contract/StoreResponse.java
@@ -8,7 +8,7 @@ public class StoreResponse {
 	private long id;
 	private String code;
 	private String name;
-	private int bailPercentage;
+	private Float bailPercentage;
 	private long appId;
 	private String csrTel;
 	private String proxyUrl;
@@ -31,10 +31,10 @@ public class StoreResponse {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getBailPercentage() {
+	public Float getBailPercentage() {
 		return bailPercentage;
 	}
-	public void setBailPercentage(int bailPercentage) {
+	public void setBailPercentage(Float bailPercentage) {
 		this.bailPercentage = bailPercentage;
 	}
 	public long getAppId() {
